fixed reRoll for fighters with tokenIds greater than 255

+++ src/FighterFarm.sol
@@ -1,11 +1,11 @@
 // SPDX-License-Identifier: MIT
 pragma solidity >=0.8.0 <0.9.0;
 
-import { FighterOps } from "./FighterOps.sol";
-import { Verification } from "./Verification.sol";
-import { AAMintPass } from "./AAMintPass.sol";
-import { AiArenaHelper } from "./AiArenaHelper.sol";
-import { Neuron } from "./Neuron.sol";
+import {FighterOps} from "./FighterOps.sol";
+import {Verification} from "./Verification.sol";
+import {AAMintPass} from "./AAMintPass.sol";
+import {AiArenaHelper} from "./AiArenaHelper.sol";
+import {Neuron} from "./Neuron.sol";
 import "@openzeppelin/contracts/token/ERC721/ERC721.sol";
 import "@openzeppelin/contracts/token/ERC721/extensions/ERC721Enumerable.sol";
 
@@ -14,7 +14,6 @@ import "@openzeppelin/contracts/token/ERC721/extensions/ERC721Enumerable.sol";
 /// @notice This contract manages the creation, ownership, and redemption of AI Arena Fighter NFTs,
 /// including the ability to mint new NFTs from a merging pool or through the redemption of mint passes.
 contract FighterFarm is ERC721, ERC721Enumerable {
-
     /*//////////////////////////////////////////////////////////////
                                 EVENTS
     //////////////////////////////////////////////////////////////*/
@@ -36,7 +35,7 @@ contract FighterFarm is ERC721, ERC721Enumerable {
     uint8[2] public maxRerollsAllowed = [3, 3];
 
     /// @notice The cost ($NRN) to reroll a fighter.
-    uint256 public rerollCost = 1000 * 10**18;    
+    uint256 public rerollCost = 1000 * 10 ** 18;
 
     /// @notice Stores the current generation for each fighter type.
     uint8[2] public generation = [0, 0];
@@ -108,7 +107,7 @@ contract FighterFarm is ERC721, ERC721Enumerable {
         _delegatedAddress = delegatedAddress;
         treasuryAddress = treasuryAddress_;
         numElements[0] = 3;
-    } 
+    }
 
     /*//////////////////////////////////////////////////////////////
                             EXTERNAL FUNCTIONS
@@ -182,7 +181,7 @@ contract FighterFarm is ERC721, ERC721Enumerable {
         _tokenURIs[tokenId] = newTokenURI;
     }
 
-    /// @notice Enables users to claim a pre-determined number of fighters. 
+    /// @notice Enables users to claim a pre-determined number of fighters.
     /// @dev The function verifies the message signature is from the delegated address.
     /// @param numToMint Array specifying the number of fighters to be claimed for each fighter type.
     /// @param signature Signature of the claim message.
@@ -193,16 +192,14 @@ contract FighterFarm is ERC721, ERC721Enumerable {
         bytes calldata signature,
         string[] calldata modelHashes,
         string[] calldata modelTypes
-    ) 
-        external 
-    {
-        bytes32 msgHash = bytes32(keccak256(abi.encode(
-            msg.sender, 
-            numToMint[0], 
-            numToMint[1],
-            nftsClaimed[msg.sender][0],
-            nftsClaimed[msg.sender][1]
-        )));
+    ) external {
+        bytes32 msgHash = bytes32(
+            keccak256(
+                abi.encode(
+                    msg.sender, numToMint[0], numToMint[1], nftsClaimed[msg.sender][0], nftsClaimed[msg.sender][1]
+                )
+            )
+        );
         require(Verification.verify(msgHash, signature, _delegatedAddress));
         uint16 totalToMint = uint16(numToMint[0] + numToMint[1]);
         require(modelHashes.length == totalToMint && modelTypes.length == totalToMint);
@@ -210,9 +207,9 @@ contract FighterFarm is ERC721, ERC721Enumerable {
         nftsClaimed[msg.sender][1] += numToMint[1];
         for (uint16 i = 0; i < totalToMint; i++) {
             _createNewFighter(
-                msg.sender, 
+                msg.sender,
                 uint256(keccak256(abi.encode(msg.sender, fighters.length))),
-                modelHashes[i], 
+                modelHashes[i],
                 modelTypes[i],
                 i < numToMint[0] ? 0 : 1,
                 0,
@@ -223,12 +220,12 @@ contract FighterFarm is ERC721, ERC721Enumerable {
 
     /// @notice Burns multiple mint passes in exchange for fighter NFTs.
     /// @dev This function requires the length of all input arrays to be equal.
-    /// @dev Each input array must correspond to the same index, i.e., the first element in each 
+    /// @dev Each input array must correspond to the same index, i.e., the first element in each
     /// array belongs to the same mint pass, and so on.
     /// @param mintpassIdsToBurn Array of mint pass IDs to be burned for each fighter to be minted.
     /// @param mintPassDnas Array of DNA strings of the mint passes to be minted as fighters.
     /// @param fighterTypes Array of fighter types corresponding to the fighters being minted.
-    /// @param modelHashes Array of ML model hashes corresponding to the fighters being minted. 
+    /// @param modelHashes Array of ML model hashes corresponding to the fighters being minted.
     /// @param modelTypes Array of ML model types corresponding to the fighters being minted.
     function redeemMintPass(
         uint256[] calldata mintpassIdsToBurn,
@@ -237,22 +234,18 @@ contract FighterFarm is ERC721, ERC721Enumerable {
         string[] calldata mintPassDnas,
         string[] calldata modelHashes,
         string[] calldata modelTypes
-    ) 
-        external 
-    {
+    ) external {
         require(
-            mintpassIdsToBurn.length == mintPassDnas.length && 
-            mintPassDnas.length == fighterTypes.length && 
-            fighterTypes.length == modelHashes.length &&
-            modelHashes.length == modelTypes.length
+            mintpassIdsToBurn.length == mintPassDnas.length && mintPassDnas.length == fighterTypes.length
+                && fighterTypes.length == modelHashes.length && modelHashes.length == modelTypes.length
         );
         for (uint16 i = 0; i < mintpassIdsToBurn.length; i++) {
             require(msg.sender == _mintpassInstance.ownerOf(mintpassIdsToBurn[i]));
             _mintpassInstance.burn(mintpassIdsToBurn[i]);
             _createNewFighter(
-                msg.sender, 
-                uint256(keccak256(abi.encode(mintPassDnas[i]))), 
-                modelHashes[i], 
+                msg.sender,
+                uint256(keccak256(abi.encode(mintPassDnas[i]))),
+                modelHashes[i],
                 modelTypes[i],
                 fighterTypes[i],
                 iconsTypes[i],
@@ -280,13 +273,7 @@ contract FighterFarm is ERC721, ERC721Enumerable {
     /// @param tokenId The ID of the fighter to update the model for.
     /// @param modelHash The hash of the machine learning model.
     /// @param modelType The type of machine learning model.
-    function updateModel(
-        uint256 tokenId, 
-        string calldata modelHash,
-        string calldata modelType
-    ) 
-        external
-    {
+    function updateModel(uint256 tokenId, string calldata modelHash, string calldata modelType) external {
         require(msg.sender == ownerOf(tokenId));
         fighters[tokenId].modelHash = modelHash;
         fighters[tokenId].modelType = modelType;
@@ -311,18 +298,16 @@ contract FighterFarm is ERC721, ERC721Enumerable {
     /// @param modelType The type of the ML model associated with the fighter.
     /// @param customAttributes Array with [element, weight] of the newly created fighter.
     function mintFromMergingPool(
-        address to, 
-        string calldata modelHash, 
-        string calldata modelType, 
+        address to,
+        string calldata modelHash,
+        string calldata modelType,
         uint256[2] calldata customAttributes
-    ) 
-        public 
-    {
+    ) public {
         require(msg.sender == _mergingPoolAddress);
         _createNewFighter(
-            to, 
-            uint256(keccak256(abi.encode(msg.sender, fighters.length))), 
-            modelHash, 
+            to,
+            uint256(keccak256(abi.encode(msg.sender, fighters.length))),
+            modelHash,
             modelType,
             0,
             0,
@@ -335,14 +320,7 @@ contract FighterFarm is ERC721, ERC721Enumerable {
     /// @param from Address of the current owner.
     /// @param to Address of the new owner.
     /// @param tokenId ID of the fighter being transferred.
-    function transferFrom(
-        address from, 
-        address to, 
-        uint256 tokenId
-    ) 
-        public 
-        override(ERC721, IERC721)
-    {
+    function transferFrom(address from, address to, uint256 tokenId) public override(ERC721, IERC721) {
         require(_ableToTransfer(tokenId, to));
         _transfer(from, to, tokenId);
     }
@@ -352,14 +330,7 @@ contract FighterFarm is ERC721, ERC721Enumerable {
     /// @param from Address of the current owner.
     /// @param to Address of the new owner.
     /// @param tokenId ID of the fighter being transferred.
-    function safeTransferFrom(
-        address from, 
-        address to, 
-        uint256 tokenId
-    ) 
-        public 
-        override(ERC721, IERC721)
-    {
+    function safeTransferFrom(address from, address to, uint256 tokenId) public override(ERC721, IERC721) {
         require(_ableToTransfer(tokenId, to));
         _safeTransfer(from, to, tokenId, "");
     }
@@ -367,7 +338,7 @@ contract FighterFarm is ERC721, ERC721Enumerable {
     /// @notice Rolls a new fighter with random traits.
     /// @param tokenId ID of the fighter being re-rolled.
     /// @param fighterType The fighter type.
-    function reRoll(uint8 tokenId, uint8 fighterType) public {
+    function reRoll(uint256 tokenId, uint8 fighterType) public {
         require(msg.sender == ownerOf(tokenId));
         require(numRerolls[tokenId] < maxRerollsAllowed[fighterType]);
         require(_neuronInstance.balanceOf(msg.sender) >= rerollCost, "Not enough NRN for reroll");
@@ -381,14 +352,11 @@ contract FighterFarm is ERC721, ERC721Enumerable {
             fighters[tokenId].element = element;
             fighters[tokenId].weight = weight;
             fighters[tokenId].physicalAttributes = _aiArenaHelperInstance.createPhysicalAttributes(
-                newDna,
-                generation[fighterType],
-                fighters[tokenId].iconsType,
-                fighters[tokenId].dendroidBool
+                newDna, generation[fighterType], fighters[tokenId].iconsType, fighters[tokenId].dendroidBool
             );
             _tokenURIs[tokenId] = "";
         }
-    }    
+    }
 
     /// @notice Returns the URI where the contract metadata is stored.
     /// @return URI where the contract metadata is stored.
@@ -407,31 +375,16 @@ contract FighterFarm is ERC721, ERC721Enumerable {
     /// @dev Calls ERC721.supportsInterface.
     /// @param _interfaceId The interface ID.
     /// @return Bool whether the interface is supported by this contract.
-    function supportsInterface(bytes4 _interfaceId)
-        public
-        view
-        override(ERC721, ERC721Enumerable)
-        returns (bool)
-    {
+    function supportsInterface(bytes4 _interfaceId) public view override(ERC721, ERC721Enumerable) returns (bool) {
         return super.supportsInterface(_interfaceId);
     }
 
     /// @notice Returns all information related to the specified fighter token ID.
     /// @param tokenId The unique identifier for the fighter token.
-    function getAllFighterInfo(
-        uint256 tokenId
-    )
+    function getAllFighterInfo(uint256 tokenId)
         public
         view
-        returns (
-            address,
-            uint256[6] memory,
-            uint256,
-            uint256,
-            string memory,
-            string memory,
-            uint16
-        )
+        returns (address, uint256[6] memory, uint256, uint256, string memory, string memory, uint16)
     {
         return FighterOps.viewFighterInfo(fighters[tokenId], ownerOf(tokenId));
     }
@@ -459,14 +412,7 @@ contract FighterFarm is ERC721, ERC721Enumerable {
     /// @param dna The dna of the fighter.
     /// @param fighterType The type of the fighter.
     /// @return Attributes of the new fighter: element, weight, and dna.
-    function _createFighterBase(
-        uint256 dna, 
-        uint8 fighterType
-    ) 
-        private 
-        view 
-        returns (uint256, uint256, uint256) 
-    {
+    function _createFighterBase(uint256 dna, uint8 fighterType) private view returns (uint256, uint256, uint256) {
         uint256 element = dna % numElements[generation[fighterType]];
         uint256 weight = dna % 31 + 65;
         uint256 newDna = fighterType == 0 ? dna : uint256(fighterType);
@@ -482,24 +428,21 @@ contract FighterFarm is ERC721, ERC721Enumerable {
     /// @param iconsType Type of icons fighter (0 means it's not an icon).
     /// @param customAttributes Array with [element, weight] of the newly created fighter.
     function _createNewFighter(
-        address to, 
-        uint256 dna, 
+        address to,
+        uint256 dna,
         string memory modelHash,
-        string memory modelType, 
+        string memory modelType,
         uint8 fighterType,
         uint8 iconsType,
         uint256[2] memory customAttributes
-    ) 
-        private 
-    {  
+    ) private {
         require(balanceOf(to) < MAX_FIGHTERS_ALLOWED);
-        uint256 element; 
+        uint256 element;
         uint256 weight;
         uint256 newDna;
         if (customAttributes[0] == 100) {
             (element, weight, newDna) = _createFighterBase(dna, fighterType);
-        }
-        else {
+        } else {
             element = customAttributes[0];
             weight = customAttributes[1];
             newDna = dna;
@@ -507,23 +450,11 @@ contract FighterFarm is ERC721, ERC721Enumerable {
         uint256 newId = fighters.length;
 
         bool dendroidBool = fighterType == 1;
-        FighterOps.FighterPhysicalAttributes memory attrs = _aiArenaHelperInstance.createPhysicalAttributes(
-            newDna,
-            generation[fighterType],
-            iconsType,
-            dendroidBool
-        );
+        FighterOps.FighterPhysicalAttributes memory attrs =
+            _aiArenaHelperInstance.createPhysicalAttributes(newDna, generation[fighterType], iconsType, dendroidBool);
         fighters.push(
             FighterOps.Fighter(
-                weight,
-                element,
-                attrs,
-                newId,
-                modelHash,
-                modelType,
-                generation[fighterType],
-                iconsType,
-                dendroidBool
+                weight, element, attrs, newId, modelHash, modelType, generation[fighterType], iconsType, dendroidBool
             )
         );
         _safeMint(to, newId);
@@ -536,11 +467,8 @@ contract FighterFarm is ERC721, ERC721Enumerable {
     /// @param tokenId The token ID of the fighter being transferred.
     /// @param to The address of the receiver.
     /// @return Bool whether the transfer is allowed or not.
-    function _ableToTransfer(uint256 tokenId, address to) private view returns(bool) {
-        return (
-          _isApprovedOrOwner(msg.sender, tokenId) &&
-          balanceOf(to) < MAX_FIGHTERS_ALLOWED &&
-          !fighterStaked[tokenId]
-        );
+    function _ableToTransfer(uint256 tokenId, address to) private view returns (bool) {
+        return
+            (_isApprovedOrOwner(msg.sender, tokenId) && balanceOf(to) < MAX_FIGHTERS_ALLOWED && !fighterStaked[tokenId]);
     }
 }
